+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # Changelog
 
-**v6.1.0 (unreleased):**
+**v6.1.0:**
 
 * Update Otter Assign to handle notebooks with an invalid kernelspec by assuming the language is Python per [#895](https://github.com/ucbds-infra/otter-grader/issues/895)
 * Fixed assignment summary in Otter Assign so that manual questions are included per [#886](https://github.com/ucbds-infra/otter-grader/issues/886)